IDs refactired to Long
Account#name added

+++ acc-20/src/main/java/net/yosifov/filipov/training/accounting/acc20/repositories/AccountsRep.java
@@ -3,5 +3,5 @@ package net.yosifov.filipov.training.accounting.acc20.repositories;
 import net.yosifov.filipov.training.accounting.acc20.entities.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountsRep  extends JpaRepository<Account, Integer> {
+public interface AccountsRep  extends JpaRepository<Account, Long> {
 }
